Fix highlighting of errors.
Syntastic by default adds a sign, which is shown in the gutter of the
buffer. Signs also add a style to the entire line, and this can be used
to style the line containing an error. However, it also shows a sign in
the gutter, which I did not want.

This commit adds a match style to the line, without requiring a sign to
be shown in the gutter.